pyproject.toml
[project]
name = "rag-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "bs4>=0.0.2",
    "langchain[anthropic]>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-text-splitters>=1.1.1",
    "langchain-google-genai>=4.2.1",
    "faiss-cpu>=1.13.2",
]
